requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
imbalanced-learn>=0.9.0

# Data Processing
pyarrow>=6.0.0
fastparquet>=0.7.0
openpyxl>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
tensorboard>=2.7.0

# Metrics and Evaluation
scikit-optimize>=0.9.0
tqdm>=4.62.0

# Configuration
pyyaml>=6.0
hydra-core>=1.1.0
omegaconf>=2.1.0

# Utilities
jupyter>=1.0.0
ipywidgets>=7.6.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code Quality
black>=21.9b0
flake8>=4.0.0
isort>=5.9.0
mypy>=0.910

# Logging and Monitoring
wandb>=0.12.0
mlflow>=1.20.0

# Additional Scientific Computing
statsmodels>=0.13.0
networkx>=2.6.0

# For RMAB and Bandit Algorithms
cvxpy>=1.1.0
cvxopt>=1.2.0

# Data Download
requests>=2.26.0
beautifulsoup4>=4.10.0
kaggle>=1.5.12

# Performance
numba>=0.54.0
joblib>=1.1.0

# Time Series Processing
tslearn>=0.5.2

# Optional: GPU acceleration helpers
# cupy-cuda11x>=9.0.0  # Uncomment for CUDA 11.x
